fix: split runner session types
