+++ src/core/runner/connection.rs
@@ -4,7 +4,7 @@ use std::thread;
 use std::time::{Duration, Instant};
 
 use chrono::Utc;
-use serde::{Deserialize, Serialize};
+use serde::Deserialize;
 use serde_json::Value;
 
 use crate::core::engine::shell;
@@ -12,128 +12,11 @@ use crate::core::error::{Error, Result};
 use crate::core::paths;
 use crate::core::server::{self, Server, ServerAuthMode, SshClient};
 
+use super::session::{
+    ReverseRunnerConnectOptions, RunnerConnectReport, RunnerDisconnectReport, RunnerFailureKind,
+    RunnerSession, RunnerSessionRole, RunnerSessionState, RunnerStatusReport, RunnerTunnelMode,
+};
 use super::{load, Runner, RunnerKind};
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(rename_all = "snake_case")]
-pub enum RunnerTunnelMode {
-    DirectSsh,
-    Reverse,
-}
-
-fn default_tunnel_mode() -> RunnerTunnelMode {
-    RunnerTunnelMode::DirectSsh
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(rename_all = "snake_case")]
-pub enum RunnerSessionRole {
-    Controller,
-    Runner,
-}
-
-fn default_session_role() -> RunnerSessionRole {
-    RunnerSessionRole::Controller
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(rename_all = "snake_case")]
-pub enum RunnerSessionState {
-    Connected,
-    Disconnected,
-    Recorded,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct RunnerSession {
-    pub runner_id: String,
-    #[serde(default = "default_tunnel_mode")]
-    pub mode: RunnerTunnelMode,
-    #[serde(default = "default_session_role")]
-    pub role: RunnerSessionRole,
-    pub server_id: Option<String>,
-    #[serde(default)]
-    pub controller_id: Option<String>,
-    #[serde(default)]
-    pub broker_url: Option<String>,
-    #[serde(default)]
-    pub remote_daemon_address: Option<String>,
-    #[serde(default)]
-    pub local_port: Option<u16>,
-    #[serde(default)]
-    pub local_url: Option<String>,
-    pub tunnel_pid: Option<u32>,
-    pub remote_daemon_pid: Option<u32>,
-    pub homeboy_version: String,
-    pub connected_at: String,
-}
-
-#[derive(Debug, Clone, Serialize, PartialEq, Eq)]
-#[serde(rename_all = "snake_case")]
-pub enum RunnerFailureKind {
-    SshFailure,
-    MissingRemoteHomeboy,
-    DaemonStartupFailure,
-    TunnelFailure,
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct RunnerConnectReport {
-    pub runner_id: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub mode: Option<RunnerTunnelMode>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub role: Option<RunnerSessionRole>,
-    pub connected: bool,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub recorded: Option<bool>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub local_url: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub broker_url: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub controller_id: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub remote_daemon_address: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub tunnel_pid: Option<u32>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub remote_daemon_pid: Option<u32>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub homeboy_version: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub session_path: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub failure_kind: Option<RunnerFailureKind>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub failure_message: Option<String>,
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct RunnerStatusReport {
-    pub runner_id: String,
-    pub connected: bool,
-    pub state: RunnerSessionState,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub session: Option<RunnerSession>,
-    pub session_path: String,
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct RunnerDisconnectReport {
-    pub runner_id: String,
-    pub disconnected: bool,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub session: Option<RunnerSession>,
-    pub session_path: String,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct ReverseRunnerConnectOptions {
-    pub controller_id: String,
-    pub runner_id: String,
-    pub broker_url: Option<String>,
-}
 
 #[derive(Debug, Clone, Deserialize)]
 struct CliEnvelope {
